fix: missing func docs and generate example in code block
