--- docs/api/hilbish/hilbish.os.md
@@ -1,20 +1,0 @@
----
-title: Module hilbish.os
-description: operating system info
-layout: doc
-menu:
-  docs:
-    parent: "API"
----
-
-## Introduction
-Provides simple text information properties about the current operating system.
-This mainly includes the name and version.
-
-## Static module fields
-|||
-|----|----|
-|family|Family name of the current OS|
-|name|Pretty name of the current OS|
-|version|Version of the current OS|
-
