+++ nature/hilbish.lua
@@ -1,3 +1,4 @@
+-- @module hilbish
 local hilbish = require 'hilbish'
 local snail = require 'snail'
 
@@ -9,9 +10,6 @@ hilbish.snail = snail.new()
 --- As a table, the caller can directly specify the standard output, error, and input
 --- streams of the command with the table keys `out`, `err`, and `input` respectively.
 --- As a boolean, it specifies whether the command should use standard output or return its output streams.
---- #param cmd string
---- #param streams table|boolean
---- #returns number, string, string
 --- #example
 --- This code is the same as `ls -l | wc -l`
 --- local fs = require 'fs'
@@ -25,6 +23,9 @@ hilbish.snail = snail.new()
 --- 	stdin = pr
 --- })
 --- #example
+-- @param cmd string
+-- @param streams table|boolean
+-- @returns number, string, string
 function hilbish.run(cmd, streams)
 	local sinks = {}
 
@@ -49,4 +50,27 @@ function hilbish.run(cmd, streams)
 	end
 
 	return table.unpack(returns)
+end
+
+--- Sets the execution/runner mode for interactive Hilbish.
+--- **NOTE: This function is deprecated and will be removed in 3.0**
+--- Use `hilbish.runner.setCurrent` instead.
+--- This determines whether Hilbish wll try to run input as Lua
+--- and/or sh or only do one of either.
+--- Accepted values for mode are hybrid (the default), hybridRev (sh first then Lua),
+--- sh, and lua. It also accepts a function, to which if it is passed one
+--- will call it to execute user input instead.
+--- Read [about runner mode](../features/runner-mode) for more information.
+-- @param mode string|function
+function hilbish.runnerMode(mode)
+	if type(mode) == 'string' then
+		hilbish.runner.setCurrent(mode)
+	elseif type(mode) == 'function' then
+		hilbish.runner.set('_', {
+			run = mode
+		})
+		hilbish.runner.setCurrent '_'
+	else
+		error('expected runner mode type to be either string or function, got', type(mode))
+	end
 end
